multimodal_rag_api/requirements.txt
fastapi
uvicorn
python-dotenv
pydantic
unstructured
langchain
langchain-openai
langchain-community
pymilvus 